fix(theme): add settings, rename note settings
